refactor(abalone) extract Play function to make it ez to make AIs
