+++ go/abalone/play.go
@@ -1,0 +1,42 @@
+package abalone
+
+import (
+	"encoding/json"
+	"fmt"
+	"log"
+	"net/http"
+
+	"github.com/danmane/abalone/go/abalone"
+	"github.com/gorilla/mux"
+)
+
+const (
+	host = ":3423"
+)
+
+type AgentInfo struct {
+	Owner  string
+	Taunts []string
+}
+
+func Play(info AgentInfo, f func(abalone.GameState) abalone.GameState) {
+	r := mux.NewRouter()
+	r.Path("/ping").HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusOK)
+		json.NewEncoder(w).Encode(&info)
+	})
+	r.Path("/move").HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		var state abalone.GameState
+		if err := json.NewDecoder(r.Body).Decode(state); err != nil {
+			w.WriteHeader(http.StatusBadRequest)
+			fmt.Fprintf(w, "error decoding game state from request body")
+			return
+		}
+
+		next := f(state)
+
+		w.WriteHeader(http.StatusOK)
+		json.NewEncoder(w).Encode(next)
+	})
+	log.Fatal(http.ListenAndServe(host, r))
+}
